Don't guard symbols with __VERSIONER_NO_GUARD
__VERSIONER_NO_GUARD has been used for the symbols that libc++ will use
it in the headers. The symbols with the macro was not guarded with
__ANDROID_API__ by the versioner.

However, this becomes a problem when the __INTRODUCED_IN macro emits the
availability attribute. Then libc++ can't use the symbols because the
clients of libc++ will experience -Wunguarded-availability if their
minSdkVersion is below the API level of the symbols annotated with
__VERSIONER_NO_GUARD.

Since __INTRODUCED_IN doesn't play a role for such symbols, we fix the
problem by not using __INTRODUCED_IN macro for the symbols with
__VERSIONER_NO_GUARD. Instead, for documenting purpose, their API levels
are written in the comment section.

Bug: 134795810
Test: m
Change-Id: Ib9db326e87df3d4730fffffdd52aa6f3b565b163