Ignore invalid TCB address.
Details:
On x86, I've noticed the traced thread-register sometimes contains a gibberish but fixed value of 0x*b (even though the ptrace() did succeed).
This probably is because the tracee is dying or does not have any TLS?

This has come up in a few LSAN/ASAN tests, causing sigsev. Ignoring it got the tests to pass, which might imply the threads  with invalid tcb aren't that interesting.

Drive-by cleanup: Fix test to check that the object is found exactly once in all dtls chunks

Change-Id: I06c86ca0c077b464fc6c9fbdf5b89889a26da5fb