Add an API for per-process disabling memory initialization.
Introduce a android_disable_mem_init() function that may be used
to disable zero- or pattern-init on non-MTE hardware, or memory
tagging on MTE hardware. The intent is that this function may
be called at any time, including when there are multiple threads
running.

Disabling zero- or pattern-init is quite trivial, we just need to set
a global variable to 0 via a Scudo API call (although there will be
some separate work required on the Scudo side to make this operation
thread-safe).

It is a bit more tricky to disable MTE across a process, because
the kernel does not provide an API for disabling tag checking in all
threads in a process, only per-thread. We need to send a signal to
each of the process's threads with a handler that issues the required
prctl call, and arrange for pthread_create() to issue the same prctl to
avoid races between thread enumeration and calls to pthread_create().

TODO:
- Write tests.
- Make the scudo side thread-safe.

Bug: 135772972
Change-Id: I81ece86ace916eb6b435ab516cd431ec4b48a3bf