Use __init_array_start+friends to find structors
Currently, we place a placeholder symbol in crtbegin.c and hope that
will end up at the beginning of each __*_array section. Unfortunately,
this conflicts with prioritized constructors, at least as implemented by
clang, as they get sorted to the front of this list.

The __init_array_start and similar symbols are injected by the linker
for exactly this use case, so we should use them instead.

This patch also fixes a minor inconsistency where dynamic init/preinit
were passed argc/argv/envp, but static were not.

Bug: 170983066
Test: mma
Change-Id: I0fffa776e5d9bdb6f8af06b4c1af148236742fef