[GWP-ASan] Provide runtime configuration through an env var + sysprop.
This patch introduces GWP_ASAN_OPTIONS="1:2:3" as well as
libc.debug.gwp_asan.<basename> (and two defaults, for system apps/system
processes, and one for non-system apps).

There are three variables that can be changed:
 1. The allocation sampling rate (default: 2500),
 2. The process sampling rate (default: 128 for system apps/processes, 1
    for opted-in apps),
 3. The number of slots available (default: 32).

If not specified, #1/#2 will return to their default values. So, a
string of "::2" means "2500:128:2". If not specified, #3 will be
calculated as a ratio of the default |2500 SampleRate : 32 slots|. So,
"1250::" will be "1250:128:64".

This change has two benefits:
 1. It allows remote tuning of GWP-ASan, through some system process
    that can modify system properties remotely, and
 2. It allows easy testing of the bionic allocator under "GWP-ASan
    torture mode", which finds bugs in GWP-ASan and bugs in libc with
    much higher accuracy.

As a result of this torture testing, this patch also fixes a couple of
minor bugs in realloc() under GWP-ASan.

Bug: 219651032
Test: atest bionic-unit-tests
Change-Id: Idb40a2a4d074e01ce3c4e635ad639a91a32d570f