Fix race when frees after main thread finishes.
When the main thread is exiting, the code disables malloc debug and
turns off all malloc debug funcionality. Unfortunately, if malloc debug
was enabled in a way that requires a header for the pointer, any frees
from other threads that occur after or during the tear down might
abort because there is an attempt to free a bad pointer. In this case,
simple leak the pointers instead of actually freeing and causing an
abort or other error.

Bug: 189541929

Test: Builds
Change-Id: I1cfe868987a8f0dc880a5b65de6709f44a5f1988