Add support for removing structs that cross blocks.
The latest kernel has a struct that needs to be removed,
but the struct crosse blocks. Add support for removing a
struct that does cross blocks. This support is very primitive,
and will not parse any really complicated struct that uses
defines in complicated ways.

Also, move the replacing of structures with a #inlcude <bits/name.h>
into removeStructs so there is a single place where structures
get removed.

Update the remove structs code in removeStructs to handle crossing
blocks, but also handles structs within structs (or unions).

Add an exit if the struct parsing does not work properly.

Add new unit tests for all of the new code.

In addition, fix the algorithm for deleting the uapi directory
before it gets updated. A new file BUILD was checking in that
directory, so delete everything in the directory except that
BUILD file.

Test: Unit tests pass.
Test: Running update_all.py results in the no unexpected changes.
Change-Id: I9a8cef0321beaf71d03b5b874327747a7edb6119