Add new purge benchmarks.
New benchmarks allocate a large chunk of memory of a certain
size, then frees a percentage of those allocations. After that
call mallopt(M_PURGE) and see how long it takes.

The idea is that this can measure the release to OS functionality
of scudo.

Also, modifies the way to parse the string arguments for benchmarks.
Instead of treating them all as a single set of range arguments for
the benchmark, treat them as a separate range 0 argument for the
benchmark.

Test: Ran benchmarks on device.
Change-Id: I76ab1b0270bb549691fda076354142f8d1d64bb6