Revert "Refactor the bp file to avoid arch variant in cc_genrule."
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/f2c6e7d6-874e-4a87-ac7d-bb61cbdf28cb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic81aede3a9b6a428509f58cf4d5c2f18f956f109