Revert "Sync with upstream NetBSD."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d588cd4e-da25-4266-8615-03caddfead5b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia20c6324a9b1b88d66dbba620478c2d8c8053ee1