Revert "Sync with upstream NetBSD."
Revert "Sync with upstream NetBSD."

This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5c2ec05f-7bf4-49e0-9cf4-f6a2e9461cbe). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I22753c8508e305bd4a2828058af64f273e0c29c6