Disable the POSIX header tests for glibc.
Our host glibc is years out of date (2.17), and will not be
updated again. We'll be using musl for the host going forward,
but glibc will fail with the next LLVM update because clang change
(https://reviews.llvm.org/D137268) stops exporting va_list.

Rather than waste time finding and backporting the glibc fix (assuming it
exists --- I couldn't find it, so it's not clear to me that ToT glibc's
<wchar.h> isn't broken by this and they just don't have a test to catch
it), let's just disable these tests on glibc...

Bug: http://b/265874089
Test: treehugger
Change-Id: Ie8b12553f4b8b971924afd8a3a0f9512297492a0