Disable the POSIX header tests for glibc.
Our glibc is years out of date, and will not be updated again. We'll be
using musl for the host going forward, but glibc will fail with the next
LLVM update because there's a clang change to not export va_list that's
presumably matched by a corresponding glibc change. Rather than waste
time finding and backporting the glibc fix, let's just disable these
tests on glibc...

Bug: http://b/265874089
Test: treehugger
Change-Id: Ie8b12553f4b8b971924afd8a3a0f9512297492a0