Revert "Verify that allocated memory is always zeroed."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d485a262-ff3c-43d5-b19e-cf1d34584d55). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6355883b41f5b61b4c66372a1726115b2cf1fc1b