Revert "Verify that allocated memory is always zeroed."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1f929b0a-4cba-40ec-81ad-56ac098c698c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1e408cf0ebb7a8af25fb61f1c1567559e5461c8a