Revert "Let clang generate most of our assembler for x86/x86-64 too."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8465fe8f-c901-48de-8d6d-2bde6c2ff1bd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I78737c7b3d2e91b13573553f8b9da59f92ea6301