Remove <sys/statvfs.h> inlines for API level 19.
API level 19 is no longer supported by the NDK.

While I'm here, let's remove the duplicated structure defintion (as
we've already done for `struct stat`), and let's take advantage of
__RENAME() to not require that you're targeting API level 21 before you
can use the (identical) *64 variants.

When a new NDK goes out that obsoletes API levels before 21, we can
either leave this as-is, or remove the __RENAME()s but bump the
__INTRODUCED_IN() back to 21 (as long as we continue to leave them
around for historical reference).

Test: treehugger
Change-Id: I9d8286f9e7ba803f3131b6dcb0486ff1b0f9d5d1