Fix mte build breakage.
The strerror_buf is way too large, so cut it down to 64 bytes. Even
this is probably too big since it is a copy of a value from
__sys_error_strings value.

Add a unit test to verify that none of the errno values are cut off
in the async_safe_format_buffer function when passing %m.

Test: New unit test passes.
Test: Changing the buffer to a small value and verify that the test fails.
Change-Id: I4cb4652709582a8a6b958e12de5d923ec950e6b6