Fix setjmp.bug_152210274 with HWASan/stack MTE enabled.
This test was flaky because there was no synchronization between the end
of the test function scope and the jumper threads, resulting in a racy
use-after-scope when accessing the thread array. Fix it by joining the
threads before leaving the test function scope.

Keep the thread array access, which acts as a kind of enforcement that
the threads are done before leaving the test, but keep the jumper threads
from crashing if they happen to get a tid of 100, by replacing the 100
with a 0 in the comparison.

Bug: 227390656
Change-Id: Icfbb0d7cae66c12e5ce31072c34529e3c5fdf563