Revert "Workaround boot loop by turning off stack protector chec..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0c8a6c99-9046-4dee-9684-d65779ff1fc4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0a1f8af2600b1095edc88a14543a1b49a71266b4