Revert "Workaround boot loop by turning off stack protector chec..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/416ff08a-9ead-49e9-aa2c-e18138590e2f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic7da43a0ab4927bccfd4aee44143bc279c35280b