Revert "Workaround boot loop by turning off stack protector chec..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2a749a50-97ab-4b06-a745-e8a379935540). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8d5cbbbcb1171c2f415d9d79eb90ffa4ceeec4e3