Revert "Workaround boot loop by turning off stack protector chec..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3ac4a1c4-12de-421f-8bdf-8d3594b76bdd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic36d4f581ed3457d6846b8756b95561967c02a47