uapi headers: simplify the <linux/compiler.h> stuff.
We don't really need <linux/compiler.h> and <linux/compiler_types.h>. We
already have a mechanism to remove unused macros, so let's do that. We
don't currently have a way to remove unused #includes, so we still need
empty <linux/compiler.h> and <linux/compiler_types.h> files (but clarify
the comments in them). We might also need to put the empty definitions
back if it turns out anyone's actually _relying_ on `__user` or
`__force` being defined but empty...

Bug: http://b/262917450
Test: treehugger
Change-Id: Iacbbbc1aeef9a4fac52dabd7811ab875cc267d4f