Support building bionic with _FILE_OFFSET_BITS==64
Support building bionic with _FILE_OFFSET_BITS=64 in preparation
for making it a default platform cflag.  Various LP32 NDK ABIs
include a 32-bit off_t, so define a __bionic_legacy_compat_off_t
that is 32-bits on LP32 even when _FILE_OFFSET_BITS==64, and
use it in all the methods that implement the old ABIs.

Bug: 275601166
Test: builds
Change-Id: Iafc3aa3de42651d4a4355cd6a705aa58b5339fb1