Add vforkrun() function.
The vfork() function has long been known to be problematic. One of its
many problems is that with MTE stack tagging it is necessary to clear
any tags written to the stack shared with the child, but the child can
exit at any time due to receiving a fatal signal, leaving the parent
not knowing how much of the stack was used by the child.

At the same time, it is the only way to spawn a child process without
using fork(), which causes the kernel to create a temporary set of page
tables with COW semantics, causing any heap writes in the original parent
to fault, even after the child has execve()'d, because the kernel makes
COW'd pages writable lazily.

Therefore, this CL adds a vforkrun() function that exposes the vfork()
functionality in a more MTE-friendly (and generally programmer-friendly)
way, by running the child on a new stack, similar to a thread. The
posix_spawn() implementation has been ported to use vforkrun().

I considered making the choice between fork() and vfork() a runtime
decision based on whether MTE stack tagging is enabled (both in vfork()
itself and in posix_spawn()), but this would diverge the behavior between
stack-MTE and non-stack-MTE builds. In theory the behavior divergence
should be unobservable modulo performance, but given that we have a number
of kernel bugs under investigation that are suspected to be triggered by
MTE and one in particular by stack MTE, it is unclear whether it would
be a good idea to add another divergence to the mix.

Bug: 274056091
Change-Id: I3721c9ae572ffb6609d365e4921455a44763750e