[HACK] Make 4k aligned APKs work with 16KB pages.
This enables loading from zip files where the uncompressed shared
libraries might not be stored on a 16k aligned offset. To do so, it
makes use of the hack for 4k aligned binaries to copy their code into
an anonymous mapping.

It comes with the same caveats as the hack for
binaries (I23d46bbfa30310b004fd026ede03095b02babad1).

Change-Id: I7ab4b6ad41163d03b0577a9730a6943a7233ba7b