riscv64: switch from x18 to gp for shadow call stack.
We want to give back a useful callee-saved general purpose register
(x18) that was only "chosen" because it was what llvm allowed for
historical reasons. gp is a better choice because it's effectively
unused otherwise anyway. Unfortunately, that means we need extra space
in jmp_buf (which I've reserved in an earlier change), so let's
rearrange the entries in jmp_buf to match their order in the register
file.

Bug: https://github.com/google/android-riscv64/issues/72
Test: treehugger
Change-Id: Ia629409a894c1a83d2052885702bbdd895c758e1