Revert "setjmp.h: add some historical detail about the _JBLEN va..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c897a5d6-8c4b-42f2-a2fe-b612aab3ea5a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6aac9770565dc64cddf07cb5298811790f0c8feb