Revert "Add versions property to libc_hwasan stubs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1d2ad1d6-dca2-4b82-92b0-68df63331ace). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iee972fbb6faf9f5074c11500e460bb89d6988a50