Revert "Nullability check for inet module."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5caf9aaf-5083-4097-99aa-64b711002692). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9bd28d0238613523f3c1c01dbc9d1334db8b79a1