Revert "Nullability check for inet module."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d6d3c4bb-7dc0-42a0-ab1b-122f11600e56). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2836fd0ac7eaa4172c00085efe7b139c4a8b1901