Revert "Nullability check for inet module."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9ac4503c-b04c-4a81-bca5-a81ca22c71ea). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I195c6b31d643905024229905d7b23bf7cb8cbdac