Revert "Nullability check for inet module."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d6962569-fb57-434f-a6a7-af77af4e6c11). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6f6435b03f45a84ba893e3694ad8cfd1cca11baa