Revert "Nullability check for if module."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c917e0de-eba2-4fc2-a83f-3b9295d4c8bc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iea677ed200681c264b17d171db2f985bfebb873a